adjust output DTO, move convertDTO logic from Food Service to Common Service, build menu data for the restaurant detail
