+++ src/dto/restaurant-detail.dto.ts
@@ -14,12 +14,8 @@ export class RestaurantDetailDTO {
   specialty: TextByLang[];
   introduction: TextByLang[];
   review_total_count: number;
-  distance_km: number;
-  delivery_time_s: number;
   cutoff_time: string[];
   having_vegeterian_food: boolean;
-  max_price: number;
-  min_price: number;
   unit: string;
   menu: FoodDTO[];
 }
