+++ src/feature/category/category.service.ts
@@ -10,6 +10,7 @@ import { FoodService } from '../food/food.service';
 import { SearchByCategory } from './dto/search-by-category-request.dto';
 import { RestaurantService } from '../restaurant/restaurant.service';
 import { PriceRange } from 'src/type';
+import { CommonService } from '../common/common.service';
 
 @Injectable()
 export class CategoryService {
@@ -20,6 +21,7 @@ export class CategoryService {
     private readonly sysCatMenuItemRepo: Repository<SysCategoryMenuItem>,
     private readonly foodService: FoodService,
     private readonly restaurantService: RestaurantService,
+    private readonly commonService: CommonService,
   ) {}
 
   async getCategories(): Promise<SysCategoryDTO[]> {
@@ -90,7 +92,7 @@ export class CategoryService {
       const restaurant = restaurants.find(
         (res) => res.restaurant_id === food.restaurant_id,
       );
-      const foodDTO = await this.foodService.convertIntoFoodDTO(
+      const foodDTO = await this.commonService.convertIntoFoodDTO(
         food,
         restaurant,
       );
