+++ src/feature/common/common.service.ts
@@ -1,19 +1,22 @@
 import { Inject, Injectable } from '@nestjs/common';
-import { RestaurantService } from '../restaurant/restaurant.service';
-import { Review } from 'src/type';
+import { DeliveryRestaurant, Review, TextByLang } from 'src/type';
 import { FlagsmithService } from 'src/dependency/flagsmith/flagsmith.service';
 import { RestaurantExt } from 'src/entity/restaurant-ext.entity';
 import { InjectEntityManager } from '@nestjs/typeorm';
-import { EntityManager } from 'typeorm';
+import { EntityManager, LessThanOrEqual, MoreThanOrEqual } from 'typeorm';
 import { TRUE } from 'src/constant';
 import { FoodRating } from 'src/entity/food-rating.entity';
+import { SkuDiscount } from 'src/entity/sku-discount.entity';
+import { SKU } from 'src/entity/sku.entity';
+import { PERCENTAGE } from 'src/constant/unit.constant';
+import { MenuItem } from 'src/entity/menu-item.entity';
+import { FoodDTO } from 'src/dto/food.dto';
 
 @Injectable()
 export class CommonService {
   constructor(
     @Inject('FLAGSMITH_SERVICE') private readonly flagService: FlagsmithService,
     @InjectEntityManager() private entityManager: EntityManager,
-    private readonly restaurantService: RestaurantService,
   ) {}
 
   async getRestaurantExtension(id: number): Promise<RestaurantExt[]> {
@@ -27,7 +30,7 @@ export class CommonService {
   }
 
   async getReviewsByRestaurantId(
-    restaurant_id: number[],
+    restaurant_id: number,
     limit: number,
     order: 'ASC' | 'DESC',
   ): Promise<Review[]> {
@@ -60,5 +63,167 @@ export class CommonService {
 
       return reviews;
     }
+  }
+
+  async getAvailableSkuDiscountBySkuId(skuId: number): Promise<SkuDiscount> {
+    const now = new Date();
+    const skuDiscount = await this.entityManager
+      .getRepository(SkuDiscount)
+      .findOne({
+        where: {
+          sku_id: skuId,
+          is_active: TRUE,
+          valid_from: LessThanOrEqual(now),
+          valid_until: MoreThanOrEqual(now),
+        },
+        order: {
+          valid_from: 'DESC',
+        },
+      });
+    return skuDiscount;
+  }
+  async getAvailableDiscountPrice(sku: SKU) {
+    let discountPrice = sku.price;
+    const skuDiscount = await this.getAvailableSkuDiscountBySkuId(sku.sku_id);
+    if (!skuDiscount) {
+      return discountPrice;
+    }
+
+    switch (skuDiscount.discount_unit_obj.symbol) {
+      case PERCENTAGE:
+        discountPrice = sku.price * (1 - skuDiscount.discount_value / 100);
+        break;
+
+      default: //VND, USD
+        discountPrice = sku.price - skuDiscount.discount_value;
+        break;
+    }
+
+    return discountPrice;
+  }
+
+  async convertIntoFoodDTO(
+    menuItem: MenuItem,
+    correspondingRestaurant: DeliveryRestaurant = null,
+  ): Promise<FoodDTO> {
+    if (this.flagService.isFeatureEnabled('fes-18-get-restaurant-detail')) {
+      //Defind Logic of Top Label later, currently just present the vegeterian sign
+      let topLabel = '';
+      if (Boolean(menuItem.is_vegetarian)) {
+        topLabel = 'CHAY';
+      }
+
+      const foodNameByLang: TextByLang[] = menuItem.menuItemExt.map((ext) => {
+        return {
+          ISO_language_code: ext.ISO_language_code,
+          text: ext.short_name,
+        };
+      });
+
+      const mainCookingMethodByLang: TextByLang[] = menuItem.menuItemExt.map(
+        (ext) => {
+          return {
+            ISO_language_code: ext.ISO_language_code,
+            text: ext.main_cooking_method,
+          };
+        },
+      );
+
+      const restaurantNameByLang: TextByLang[] =
+        correspondingRestaurant?.restaurant_ext.map((ext) => {
+          return {
+            ISO_language_code: ext.ISO_language_code,
+            text: ext.name,
+          };
+        }) || null;
+
+      return {
+        id: menuItem.menu_item_id,
+        image: menuItem.image_obj.url,
+        top_label: topLabel,
+        bottom_label: null,
+        name: foodNameByLang,
+        restaurant_name: restaurantNameByLang,
+        restaurant_id: menuItem.restaurant_id,
+        calorie_kcal: menuItem.skus[0].calorie_kcal,
+        rating: menuItem.rating,
+        distance_km: correspondingRestaurant?.distance_km || null,
+        delivery_time_s: correspondingRestaurant?.delivery_time_s || null,
+        main_cooking_method: mainCookingMethodByLang,
+        ingredient_brief_vie: menuItem.ingredient_brief_vie,
+        ingredient_brief_eng: menuItem.ingredient_brief_eng,
+        price: menuItem.skus[0].price,
+        price_after_discount: await this.getAvailableDiscountPrice(
+          menuItem.skus[0],
+        ),
+        promotion: menuItem.promotion,
+        cutoff_time: menuItem.cutoff_time,
+        preparing_time_s: menuItem.preparing_time_s,
+        cooking_time_s: menuItem.cooking_time_s,
+        quantity_available: menuItem.quantity_available,
+        is_vegetarian: Boolean(menuItem.is_vegetarian),
+        cooking_schedule: menuItem.cooking_schedule,
+        units_sold: menuItem.units_sold,
+      };
+    }
+    //CURRENT LOGIC
+    //Defind Logic of Top Label later, currently just present the vegeterian sign
+    let topLabel = '';
+    if (Boolean(menuItem.is_vegetarian)) {
+      topLabel = 'CHAY';
+    }
+
+    const foodNameByLang: TextByLang[] = menuItem.menuItemExt.map((ext) => {
+      return {
+        ISO_language_code: ext.ISO_language_code,
+        text: ext.short_name,
+      };
+    });
+
+    const mainCookingMethodByLang: TextByLang[] = menuItem.menuItemExt.map(
+      (ext) => {
+        return {
+          ISO_language_code: ext.ISO_language_code,
+          text: ext.main_cooking_method,
+        };
+      },
+    );
+
+    const restaurantNameByLang: TextByLang[] =
+      correspondingRestaurant.restaurant_ext.map((ext) => {
+        return {
+          ISO_language_code: ext.ISO_language_code,
+          text: ext.name,
+        };
+      });
+
+    return {
+      id: menuItem.menu_item_id,
+      image: menuItem.image_obj.url,
+      top_label: topLabel,
+      bottom_label: null,
+      name: foodNameByLang,
+      restaurant_name: restaurantNameByLang,
+      restaurant_id: menuItem.restaurant_id,
+      calorie_kcal: menuItem.skus[0].calorie_kcal,
+      rating: menuItem.rating,
+      distance_km: correspondingRestaurant.distance_km,
+      delivery_time_s: correspondingRestaurant.delivery_time_s,
+      main_cooking_method: mainCookingMethodByLang,
+      ingredient_brief_vie: menuItem.ingredient_brief_vie,
+      ingredient_brief_eng: menuItem.ingredient_brief_eng,
+      price: menuItem.skus[0].price,
+      price_after_discount: await this.getAvailableDiscountPrice(
+        menuItem.skus[0],
+      ),
+      promotion: menuItem.promotion,
+      cutoff_time: menuItem.cutoff_time,
+      preparing_time_s: menuItem.preparing_time_s,
+      cooking_time_s: menuItem.cooking_time_s,
+      quantity_available: menuItem.quantity_available,
+      is_vegetarian: Boolean(menuItem.is_vegetarian),
+      cooking_schedule: menuItem.cooking_schedule,
+      units_sold: menuItem.units_sold,
+    };
   }
 }
