+++ src/feature/restaurant/restaurant.service.ts
@@ -9,6 +9,10 @@ import { FlagsmithService } from 'src/dependency/flagsmith/flagsmith.service';
 import { GeneralResponse } from 'src/dto/general-response.dto';
 import { RestaurantDetailDTO } from 'src/dto/restaurant-detail.dto';
 import { Media } from 'src/entity/media.entity';
+import { CommonService } from '../common/common.service';
+import { RestaurantExt } from 'src/entity/restaurant-ext.entity';
+import { FoodDTO } from 'src/dto/food.dto';
+import { TRUE } from 'src/constant';
 
 @Injectable()
 export class RestaurantService {
@@ -18,6 +22,7 @@ export class RestaurantService {
     private restaurantRepo: Repository<Restaurant>,
     private readonly ahamoveService: AhamoveService,
     @InjectEntityManager() private entityManager: EntityManager,
+    private readonly commonService: CommonService,
   ) {}
 
   async findAll(): Promise<Restaurant[]> {
@@ -185,9 +190,23 @@ export class RestaurantService {
         .leftJoinAndSelect('restaurant.logo', 'logo')
         .leftJoinAndSelect('restaurant.unit_obj', 'unit')
         .leftJoinAndSelect('restaurant.address', 'address')
+        .leftJoinAndSelect('restaurant.menu_items', 'menuItem')
+        .leftJoinAndSelect('menuItem.menuItemExt', 'menuItemExt')
+        .leftJoinAndSelect('menuItem.image_obj', 'image')
+        .leftJoinAndSelect('menuItem.skus', 'sku')
         .where('restaurant.restaurant_id = :restaurant_id', { restaurant_id })
         .andWhere('restaurant.is_active = 1')
+        .andWhere('sku.is_standard = :standard', {
+          standard: TRUE,
+        })
+        .andWhere('sku.is_active = :active', { active: TRUE })
         .getOne();
+
+      if (!restaurant) {
+        result.statusCode = 404;
+        result.message = 'Restaurant not found';
+        return result;
+      }
 
       //Get medias
       const medias = await this.getAllMediaByRestaurantId(restaurant_id);
@@ -196,6 +215,33 @@ export class RestaurantService {
       const restaurant_ext = await this.convertRestaurantExtToTextByLang(
         restaurant.restaurant_ext,
       );
+
+      //Get reviews
+      const reviews = await this.commonService.getReviewsByRestaurantId(
+        restaurant_id,
+        5,
+        'DESC',
+      );
+
+      //Get menu
+      const menuItems = await restaurant.menu_items;
+      const convertedMenuItems: FoodDTO[] = [];
+      let having_vegeterian_food: boolean = false;
+      const cutoff_time = [];
+      for (const menuItem of menuItems) {
+        //Check if having vegeterian food
+        if (menuItem.is_vegetarian === 1) {
+          having_vegeterian_food = true;
+        }
+
+        //push cutoff time
+        cutoff_time.push(menuItem.cutoff_time);
+
+        //Convert to FoodDTO
+        const foodDTO = await this.commonService.convertIntoFoodDTO(menuItem);
+        convertedMenuItems.push(foodDTO);
+      }
+      cutoff_time.sort();
 
       //Mapping data with output
       const data: RestaurantDetailDTO = {
@@ -219,19 +265,15 @@ export class RestaurantService {
         rating: restaurant.rating,
         top_food: restaurant.top_food,
         promotion: restaurant.promotion,
-        reviews: [],
+        reviews: reviews,
         name: restaurant_ext.name,
         specialty: restaurant_ext.specialty,
         introduction: restaurant_ext.introduction,
         review_total_count: restaurant.review_total_count,
-        distance_km: null,
-        delivery_time_s: null,
-        cutoff_time: [],
-        having_vegeterian_food: null,
-        max_price: null,
-        min_price: null,
+        cutoff_time: cutoff_time,
+        having_vegeterian_food: having_vegeterian_food,
         unit: restaurant.unit_obj.symbol,
-        menu: [],
+        menu: convertedMenuItems,
       };
 
       result.statusCode = 200;
